*: login to the game, fix some memory corruption and re-implement TileRenderer (#49)

+++ Client/Packet/Types/SystemTypes.h
@@ -27,6 +27,7 @@
 #include <iostream>
 #include <fstream>
 #include <string>
+#include <cstdint>
 
 //using namespace std;
 using std::string;
@@ -67,11 +68,11 @@ const unsigned int szbool   = sizeof(bool);
 const unsigned int szchar   = sizeof(char);
 const unsigned int szshort  = sizeof(short);
 const unsigned int szint    = sizeof(int);
-const unsigned int szlong   = sizeof(long);
+const unsigned int szlong   = sizeof(int32_t);
 const unsigned int szuchar  = sizeof(unsigned char);
 const unsigned int szushort = sizeof(unsigned short);
 const unsigned int szuint   = sizeof(unsigned int);
-const unsigned int szulong  = sizeof(unsigned long);
+const unsigned int szulong  = sizeof(uint32_t);
 const unsigned int szBYTE   = sizeof(BYTE);
 const unsigned int szWORD   = sizeof(WORD);
 const unsigned int szDWORD  = sizeof(DWORD);
@@ -112,7 +113,7 @@ enum WorldStatus
 	WORLD_CLOSE
 };
 
-typedef unsigned long IP_t;
+typedef uint32_t IP_t;
 const uint szIP = sizeof(IP_t);
 
 typedef BYTE WorldID_t;
